// client/src/enums/locality.enum.ts
export enum LocalityEnum {
  Baryshivka = "Баришівка",
  BilaTserkva = "Біла Церква",
  Bohuslav = "Богуслав",
  Borodianka = "Бородянка",
  Brovary = "Бровари",
  Vasylkiv = "Васильків",
  Vyshhorod = "Вишгород",
  Hlevakha = "Глеваха",
  Hrebinky = "Гребінки",
  Ivankiv = "Іванків",
  Irpin = "Ірпінь",
  Kaharlyk = "Кагарлик",
  Kyiv = "Київ",
  Makariv = "Макарів",
  Mironivka = "Миронівка",
  Obukhiv = "Обухів",
  PereiaslavKhmelnytskyi = "Переяслав-Хмельницький",
  Rzhyschiv = "Ржищів",
  Skvyra = "Сквира",
  Slavutych = "Славутич",
  Tarashcha = "Тараща",
  Tetiiv = "Тетіїв",
  Ukrayinka = "Українка",
  Vasylivka = "Василівка",
  Yahotyn = "Яготин"
}